Flask/requirements.txt
gunicorn
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.9.0
Flask-FontAwesome==0.1.5
flask-cors
numpy==1.16.0
scipy==1.5.2
python-dotenv==0.15.0
pandas==1.1.3
h5py==2.10.0
tensorflow==2.3.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
tqdm==4.48.2
Pillow==7.2.0
